Move tests and add more coverage
